+++ NEWS.md
@@ -1,5 +1,7 @@
 # Elbird 0.0.1
 
+* Fix add_user_word function with cache.
+
 # Elbird 0.0.0.9001
 
 * Added a `NEWS.md` file to track changes to the package.